Move streetview tasks back to streetview task file.
