+++ geoapi/tasks/__init__.py
@@ -1,3 +1,3 @@
 from geoapi.tasks.lidar import convert_to_potree
-from geoapi.tasks.external_data import import_file_from_agave, import_from_agave, refresh_observable_projects, process_streetview_sequences
-from geoapi.tasks.streetview import publish, from_tapis_to_streetview
+from geoapi.tasks.external_data import import_file_from_agave, import_from_agave, refresh_observable_projects
+from geoapi.tasks.streetview import publish, from_tapis_to_streetview, process_streetview_sequences
